[#20] move config file in config package
