+++ examples/sample2/scene/config/screen_size.go
@@ -1,4 +1,4 @@
-package scene
+package config
 
 const (
 	SCREEN_WIDTH  = 960
